requirements.txt
attrs                25.3.0
certifi              2025.6.15
cffi                 1.17.1
colorama             0.4.6
core-universal4      4.38.0
eyes-common          6.3.1
eyes-selenium        6.2.2
h11                  0.16.0
idna                 3.10
iniconfig            2.1.0
marshmallow          3.26.1
outcome              1.3.0.post0
packaging            25.0
pip                  25.1.1
pluggy               1.6.0
pycparser            2.22
Pygments             2.19.2
PySocks              1.7.1
pytest               8.4.1
python-dotenv        1.1.1
selenium             4.34.1
sniffio              1.3.1
sortedcontainers     2.4.0
trio                 0.30.0
trio-websocket       0.12.2
typing_extensions    4.14.1
urllib3              2.5.0
websocket-client     1.8.0
wsproto              1.2.0